create central state store
